% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCBF-deprecated.R, R/entropy.R
\name{FCBF-deprecated}
\alias{FCBF-deprecated}
\alias{SU}
\alias{IG}
\title{FCBF-deprecated.r
 @title Deprecated functions in package \pkg{FCBF}.
 @description The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.}
\usage{
SU(x, y, base = 2)

IG(x, y, base = 2)
}
\description{
FCBF-deprecated.r
 @title Deprecated functions in package \pkg{FCBF}.
 @description The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\section{\code{SU}}{

For \code{SU}, use \code{\link{get_SU_for_vector_pair}}.
}

\section{\code{IG}}{

For \code{IG}, use \code{\link{get_IG_for_vector_pair}}.
}

\keyword{internal}
