% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_modules.R
\name{find_modules}
\alias{find_modules}
\title{find_modules}
\usage{
find_modules(discretized_exprs, target, FCBF_threshold = 0.1,
  verbose = FALSE)
}
\arguments{
\item{discretized_exprs}{A table of features (samples in rows, variables in columns, and each observation in each cell)}

\item{target}{A target vector, factor containing classes of the observations. Note: the
observations must be in the same order as the parameter x}

\item{FCBF_threshold}{A threshold for the minimum correlation (as determined by symettrical uncertainty)
between each variable and the class used for wrapped FCBF function. Defaults to 0.1.}

\item{verbose}{Adds verbosity. Defaults to FALSE.}
}
\value{
Returns a list with the CBF modules found
}
\description{
find_modules uses Symmetrical Uncertainty as a correlation measure and the FCBF algorithm to
}
\details{
1 - Filter the gene list by correlations to a class (Step 1)

and

2 - Determine soft thresholds for coexpression to genes predominantly correlated to a class.
}
\examples{
 data(scDengue)
 exprs <- SummarizedExperiment::assay(scDengue, 'logcounts')
 discrete_expression <- as.data.frame(discretize_exprs(exprs))
 head(discrete_expression[,1:4])
 infection <- SummarizedExperiment::colData(scDengue)
 target <- infection$infection
 cbf_modules <- find_modules(discrete_expression,target, FCBF_threshold = 0.06)
}
