% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_of_features.R
\name{get_su-deprecated}
\alias{get_su-deprecated}
\title{Symmetrical Uncertainty diagnostic}
\arguments{
\item{x}{A table of features (observations in rows, variables in columns)}

\item{y}{A target vector, factor containing classes of the observations. Note: the
observations must be in the same order as the parameter x.}

\item{samples_in_rows}{A flag for the case in which samples are in rows and variables/genes in columns. Defaults to FALSE.}

\item{bar_of_progress}{A flag to show progress. Defaults to FALSE.}
}
\value{
A dataframe containing the SU values for each feature
}
\description{
This functions runs symmetrical uncertainty for a feature table and a class, returning
the scores of symmetrical uncertainty for all features
}
\seealso{
\code{\link{FCBF-deprecated}}
}
\keyword{internal}
