% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_of_features.R
\name{get_su_for_feature_table_and_vector}
\alias{get_su_for_feature_table_and_vector}
\title{Symmetrical Uncertainty diagnostic}
\usage{
get_su_for_feature_table_and_vector(
  feature_table,
  target_vector,
  samples_in_rows = FALSE,
  bar_of_progress = FALSE
)
}
\arguments{
\item{feature_table}{A table of features (observations in rows, variables in columns)}

\item{target_vector}{A target vector, factor containing classes of the observations. Note: the
observations must be in the same order as the parameter x.}

\item{samples_in_rows}{A flag for the case in which samples are in rows and variables/genes in columns. Defaults to FALSE.}

\item{bar_of_progress}{A flag to show progress. Defaults to FALSE.}
}
\value{
A dataframe containing the SU values for each feature
}
\description{
This functions runs symmetrical uncertainty for a feature table and a class, returning
the scores of symmetrical uncertainty for all features
}
\examples{
data(scDengue)
exprs <- SummarizedExperiment::assay(scDengue, 'logcounts')
discrete_expression <- as.data.frame(discretize_exprs(exprs))
infection <- SummarizedExperiment::colData(scDengue)
target <- infection$infection
su_values <- get_su_for_feature_table_and_vector(discrete_expression[,],target[])
su_values[1:10,]
}
