% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_SU_for_vector_pair}
\alias{get_SU_for_vector_pair}
\title{get_SU_for_vector_pair
 Formula for symetrical uncertainty as described in  Yu, L. and Liu, H. , 2003.
This functions runs symmetrical uncertainty for two features,
returning the score}
\usage{
get_SU_for_vector_pair(x, y, base = 2)
}
\arguments{
\item{x}{A vector containing a categorical feature}

\item{y}{A vector containing other categorical feature}

\item{base}{The base used for the logaritmic function. The default is exp(1) (~2.718)}
}
\value{
A numerical value for the Symetrical Uncertainty score
}
\description{
get_SU_for_vector_pair
 Formula for symetrical uncertainty as described in  Yu, L. and Liu, H. , 2003.
This functions runs symmetrical uncertainty for two features,
returning the score
}
\examples{
 data(scDengue)
 exprs <- SummarizedExperiment::assay(scDengue, 'logcounts')
 discrete_expression <- as.data.frame(discretize_exprs(exprs))
 discrete_expression_gene_1 <- discrete_expression$V1
 discrete_expression_gene_2 <- discrete_expression$V2
 SU(discrete_expression_gene_1,discrete_expression_gene_2)
}
