% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{IG-deprecated}
\alias{IG-deprecated}
\title{Information Gain
This functions runs Information Gain for two features,
returning the score}
\arguments{
\item{x}{A vector containing a categorical feature}

\item{y}{A vector containing other categorical feature}

\item{base}{The base used for the logaritmic function. The default is exp(1) (~2.718)}
}
\value{
A numerical value for the Information Gain score
}
\description{
Information Gain
This functions runs Information Gain for two features,
returning the score
}
\seealso{
\code{\link{FCBF-deprecated}}
}
\keyword{internal}
