% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{IG}
\alias{IG}
\title{Information Gain
This functions runs Information Gain for two features,
returning the score}
\usage{
IG(x, y, base = 2)
}
\arguments{
\item{x}{A vector containing a categorical feature}

\item{y}{A vector containing other categorical feature}

\item{base}{The base used for the logaritmic function. The default is exp(1) (~2.718)}
}
\value{
A numerical value for the Information Gain score
}
\description{
Information Gain
This functions runs Information Gain for two features,
returning the score
}
\examples{
  data(scDengue)
  exprs <- SummarizedExperiment::assay(scDengue, 'logcounts')
  discrete_expression <- as.data.frame(discretize_exprs(exprs))
  discrete_expression_gene_1 <- discrete_expression$V1
  discrete_expression_gene_2 <- discrete_expression$V2
  IG(discrete_expression_gene_1,discrete_expression_gene_2)
}
