% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ig.R
\docType{data}
\name{ig_values}
\alias{ig_values}
\title{Get information gain}
\format{An object of class \code{data.frame} with 19541 rows and 1 columns.}
\usage{
ig_values
}
\arguments{
\item{x}{A table of features (observations in rows, variables in columns)}

\item{y}{A target vector, factor containing classes of the observations. Note: the
observations must be in the same order as the parameter x.}
}
\value{
A dataframe containing the SU values for each feature
}
\description{
This functions runs information gain for a feature table and a class, returning
the scores of information gain for all features
}
\examples{

data(scDengue)
exprs <- SummarizedExperiment::assay(scDengue, 'logcounts')
discrete_expression <- as.data.frame(discretize_exprs(exprs))
infection <- SummarizedExperiment::colData(scDengue)
target <- infection$infection
}
\keyword{datasets}
