% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{plot_one_module}
\alias{plot_one_module}
\title{plot_one_module}
\usage{
plot_one_module(adjacency_matrix, name, color = "#B30000FF", n = 10)
}
\arguments{
\item{adjacency_matrix}{An adjacency matrix from the find modules function
(dataframe with a 'genes' column)}

\item{name}{The name of the plot.}

\item{color}{The color of the plot. Defaults to a shade of red ("#B30000FF").}

\item{n}{The max number of gene labels to show.  Defaults to 10.}

\item{genes}{A character vector with the FCBF selected genes}
}
\value{
A gg object
}
\description{
Takes as input an adjacency matrix from the find_modules function
(output = "matrix") an a list of hubs (FCBF - selected)
(Networks are inspired in the CEMiTool package networks)
And returns a pdf plotting the modules with their connections
}
