% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{scDengue}
\alias{scDengue}
\title{Dengue infected macrophages; gene expression data from GEO study GSE110496}
\format{An object of class \code{SingleCellExperiment}}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE110496}{GEO}
}
\usage{
data(scDengue)
}
\description{
Expression data from single cells, from  adengue virus infection study by
Zanini et al, #' 2018. The expression was filtered to get cells 12 hours
after infection with #' a multiplicity of infection (moi) of 1 (dengue)
or uninfected(ctrl). Gene counts were normalized via Bioconductor
package "SCNorm".
}
\details{
Gene expression has to be discretized for use in FCBF.
}
\examples{
data(scDengue)
data(scDengue)
exprs <- SummarizedExperiment::assay(scDengue, 'logcounts')
infection <- SummarizedExperiment::colData(scDengue)
target <- infection$infection
# Discretize gene expression
discrete_expression <- as.data.frame(discretize_exprs(exprs))
fcbf_features <- fcbf(discrete_expression,
                              target,
                              thresh = 0.05,
                              verbose = TRUE)
}
\references{
Zanini, F., Pu, S. Y., Bekerman, E., Einav, S., & Quake, S. R.
(2018). Single-cell transcriptional dynamics of flavivirus infection.
Elife, 7, e32942.
\href{https://www.ncbi.nlm.nih.gov/pubmed/29451494}{PubMed}
}
\keyword{datasets,}
\keyword{dengue,}
\keyword{single-cell}
