% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_plot.R
\name{su_plot}
\alias{su_plot}
\title{Symmetrical Uncertainty diagnostic}
\usage{
su_plot(x, y)
}
\arguments{
\item{x}{A table of features (observations in rows, variables in columns)}

\item{y}{A target vector, factor containing classes of the observations. Note: the
observations must be in the same order as the parameter x.}
}
\value{
Plots an histogram of symmetrical uncertainty values regarding the class.
}
\description{
This functions runs symmetrical uncertainty for a feature table and a class, returning
an histogram of the scores
}
\examples{
data(scDengue)
exprs <- SummarizedExperiment::assay(scDengue, 'logcounts')
discrete_expression <- as.data.frame(discretize_exprs(exprs))
infection <- SummarizedExperiment::colData(scDengue)
target <- infection@listData$infection
su_plot(discrete_expression,target)
}
