% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_exprs.R
\name{discretize_exprs}
\alias{discretize_exprs}
\title{Simple discretizing of gene expression}
\usage{
discretize_exprs(expression_table, number_of_bins = 3)
}
\arguments{
\item{expression_table}{A previously normalized expression table (genes/variables in rows,
samples/observations in columns)
Note: this might drastically change the number of selected features.}

\item{number_of_bins}{Number of equal-width bins for discretization.
Note: it is a binary discretization, with the
first bin becoming one class ('low') and the other bins, another class ('high').
Defaults to 3.}
}
\value{
A data frame with the discretized features (observations in rows, variables in columns)
}
\description{
This function takes the range of values for each gene in a
previously normalized expression table (genes/variables in rows,
samples/observations in columns) and uses it for a width-based discretization.
Each feature is divide into "n" bins of equal width. The first bin is
attributed the class 'low' and the next bins are assigned to "high".
}
\examples{
data(single_cell_dengue_exprs)
discrete_expression <- as.data.frame(discretize_exprs(single_cell_dengue_exprs))
head(discrete_expression[,1:4])
}
