% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_exprs_supervised.R
\name{discretize_exprs_supervised}
\alias{discretize_exprs_supervised}
\title{supervised_disc_df}
\usage{
discretize_exprs_supervised(expression_table, target, parallel = FALSE)
}
\arguments{
\item{expression_table}{A previously normalized expression table}

\item{target}{A series of labels matching each of the values in the gene vector
(genes in rows, cells/samples in columns)}

\item{parallel}{Set calculations in parallel. May be worth it if the number of rows and columns is really large. Do watchout for memory overload.}
}
\value{
A data frame with the discretized features in the same order as previously
}
\description{
Uses several discretizations and selects the one that is best for a given variable (gene)
in comparison to a target class by equivocation
}
\examples{
data(scDengue)
exprs <- as.data.frame(SummarizedExperiment::assay(scDengue, 'logcounts'))
exprs <- exprs [1:200, 1:120]
infection <- SummarizedExperiment::colData(scDengue)
target <- infection$infection
discrete_expression <- as.data.frame(discretize_exprs_supervised(exprs,target))
fcbf(discrete_expression,target, minimum_su = 0.05, verbose = TRUE)
}
