% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCBF-deprecated.R, R/entropy.R,
%   R/ranking_of_features.R
\name{FCBF-deprecated}
\alias{FCBF-deprecated}
\alias{SU}
\alias{IG}
\alias{get_su}
\alias{get_ig}
\title{FCBF-deprecated.r
 @title Deprecated functions in package \pkg{FCBF}.
 @description The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.}
\usage{
SU(x, y, base = 2)

IG(x, y, base = 2)

get_su(x, y)

get_ig(x, y)
}
\arguments{
\item{x}{A table of features (observations in rows, variables in columns)}

\item{y}{A target vector, factor containing classes of the observations. Note: the
observations must be in the same order as the parameter x.}
}
\value{
A dataframe containing the SU values for each feature
}
\description{
This functions runs information gain for a feature table and a class, returning
the scores of information gain for all features
}
\section{\code{SU}}{

For \code{SU}, use \code{\link{get_SU_for_vector_pair}}.
}

\section{\code{IG}}{

For \code{IG}, use \code{\link{get_IG_for_vector_pair}}.
}

\section{\code{get_su}}{

For \code{get_su}, use \code{\link{get_su_for_feature_table_and_vector}}.


For \code{get_su}, use \code{\link{get_su_for_feature_table_and_vector}}.
}

\keyword{internal}
