% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcbf.R
\name{fcbf}
\alias{fcbf}
\title{Fast Correlation Based Filter function.}
\usage{
fcbf(x, y, thresh = 0.25, verbose = FALSE)
}
\arguments{
\item{x}{A table of features (samples in rows, variables in columns, and each observation in each cell)}

\item{y}{A target vector, factor containing classes of the observations. Note: the
observations must be in the same order as the parameter x}

\item{thresh}{A threshold for the minimum correlation (as determined by symettrical uncertainty)
between each variable and the class. Defaults to 0.25.
Note: this might drastically change the number of selected features.}

\item{verbose}{Adds verbosity. Defaults to FALSE.}
}
\value{
Returns a data frame with the selected features index (first row) and their symmetrical uncertainty values regarding the class (second row). Variable names are present in rownames
}
\description{
This functions allows selection of variables from a feature table
of discrete/categorial variables and a target class.
The function is based on the algorithm described in
Yu, L. and Liu, H.; Feature Selection
for High-Dimensional Data A Fast Correlation Based Filter Solution,
Proc. 20th Intl. Conf. Mach. Learn. (ICML-2003), Washington DC, 2003
}
\details{
Obs: For gene expression, you will need to run discretize_exprs first
}
\examples{
data(single_cell_dengue_exprs)
discrete_expression <- as.data.frame(discretize_exprs(single_cell_dengue_exprs))
head(discrete_expression[,1:4])
data("single_cell_dengue_annot")
fcbf(discrete_expression,target, tresh = 0.05)
}
